patched the renderer, results differ from previous experiments, reason unknown

+++ fibo_render/src/constants.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "graphic")]
-pub const MOVE_STEP: u64 = 100;
+pub const MOVE_STEP: usize = 100;
 
 #[cfg(feature = "graphic")]
-pub const SHOW_IMAGE_TIMES: u32 = 5;
+pub const SHOW_IMAGE_TIMES: usize = 5;
 
 pub const PROGRESS_BAR_SIZE: u32 = 50;
